Further improvements
Add utils file. For now with a simple function to replace a rune
that causes an encoding to fail for one that doesn't. (more will
be added to this process).

Add gitignore and remove testdir from files tracked.

Update response from acoustid to include sources and use this value
to sort the results so its easier to find the (most likely to be)
correct match.

Small fixes.